+++ Runtime/Interaction/Controllers/ActionBasedController.cs
@@ -235,7 +235,7 @@ namespace UnityEngine.XR.Interaction.Toolkit
             if (action == null)
                 return false;
 
-#if INPUT_SYSTEM_1_1_OR_NEWER
+#if INPUT_SYSTEM_1_1_OR_NEWER || INPUT_SYSTEM_1_1_PREVIEW // 1.1.0-preview.2 or newer, including pre-release
                 return action.phase == InputActionPhase.Performed;
 #else
             if (action.activeControl is ButtonControl buttonControl)